requirements.txt
requests~=2.26.0
selenium~=3.141.0
pytesseract~=0.3.8
Pillow~=8.1.2